requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
berserk==0.13.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
databases==0.9.0
Deprecated==1.2.14
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
joblib==1.4.2
load-dotenv==0.1.0
Mako==1.3.5
MarkupSafe==2.1.5
ndjson==0.3.1
numpy==2.1.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.18
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.0
sudo==1.0.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
wrapt==1.16.0
xgboost==2.1.1

setuptools~=75.5.0
stockfish~=3.28.0